# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Spongeling Admin API
  version: 1.0.0
  description: API to add pos and save patterns
servers:
  - url: '{protocol}://{server}'
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
      server:
        default: localhost:6543
security:
  - BasicAuth: []
...
paths:
...
  /word/class:
    get:
      tags:
        - Classification
      operationId: get-all-custom-pos
      summary: Endpoint to get all custom pos
      description: A web service endpoint for fetching all custom pos page-by-page with description and associated words
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseGetAllWordClass'
          description: Valid response
    post:
      tags:
        - Classification
      operationId: add-custom-pos
      summary: Endpoint to save a custom pos
      description: A web service endpoint for saving a custom pos with description and associated words
      requestBody:
        description: Request object contaning custom pos information
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestAddWordClass'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseAddWordClass'
          description: Valid response
  /word/class/{class_id}:
    patch:
      tags:
        - Classification
      operationId: update-custom-pos
      summary: Endpoint to update a custom pos
      description: A web service endpoint for updating a custom pos and it's description and associated words against the custom pos identifier
      parameters:
        - deprecated: false
          name: class_id
          description: custom pos identifier
          schema:
            type: integer
          in: path
          required: true
      requestBody:
        description: Request object contaning custom pos information
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestUpdateWordClass'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseUpdateWordClass'
          description: Valid response
    delete:
      tags:
        - Classification
      operationId: delete-custom-pos
      summary: Endpoint to delete a custom pos
      description: A web service endpoint for deleting a custom pos against the it's identifier
      parameters:
        - deprecated: false
          name: class_id
          description: custom pos identifier
          schema:
            type: integer
          in: path
          required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseDeleteWordClass'
          description: Valid response
components:
  schemas:
...
    ResponseGetAllWordClass:
      type: array
      items:
        $ref: '#/components/schemas/WordClass'
    RequestAddWordClass:
      $ref: '#/components/schemas/WordClass'
    ResponseAddWordClass:
      $ref: '#/components/schemas/Response'
    RequestUpdateWordClass:
      $ref: '#/components/schemas/WordClass'
    ResponseUpdateWordClass:
      $ref: '#/components/schemas/Response'
    ResponseDeleteWordClass:
      $ref: '#/components/schemas/Response'
      
    # PATTERNS
...
    Words:
      required:
        - words
      properties:
        words:
          type: array
          items:
            type: string
      example:
        words:
          - blanco
          - negro
          - azul
          - rojo
    WordClass:
      required:
        - name
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        words:
          $ref: '#/components/schemas/Words'
      example:
        name: COL
        description: List of colours to be used to show how they are used to describe things.
        words:
          - blanco
          - negro
          - azul
          - rojo
          
    # Shared
    Response:
      type: object
      properties:
        id:
          type: integer
        status:
          type: integer
        message:
          type: string
      example:
        id: 17
        status: 200
        message: successfully done
  securitySchemes:
    BasicAuth:     # <-- arbitrary name for the security scheme
      type: http
      scheme: basic